requirements.txt
opencv-python
numpy
google-generativeai
python-dotenv
Flask
